src/job-market-gov/pyproject.toml
[project]
name = "job-market-gov"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.42.1",
]
